soccer-streaming/requirements.txt
kafka-python==2.0.2
boto3==1.34.0
dataclasses-json==0.6.3
python-dotenv==1.0.0